Make Category a plain old Crystal object
Categories are needed to render every page because they are used to
produce the list of feeds in the <head>. Since they will change very
rarely and there is no way to create them in the UI to commit changes
them to be static data.

This also eliminates the need to seed the database with required data
and the special AllCategory class.